% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{copepod}
\alias{copepod}
\title{Copepod Data}
\format{
A data frame with 43 rows and 5 variables:
\describe{
\item{stations}{A character vector of station identifiers}
\item{year}{A numeric vector containing year}
\item{taxa}{A character vector indicating taxa type}
\item{species}{A character vector indicating species type}
\item{abundance}{A numeric vector of abundances}
}
}
\usage{
copepod
}
\description{
A dataset containing stations, auxiliary information, and
abundance. Because we could not share the raw data, we provided this
data set as a means to emulate the raw data.
}
\keyword{datasets}
