% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.correlation.R
\name{sentinel.correlation}
\alias{sentinel.correlation}
\title{Compute sentinel station statistics}
\usage{
\method{sentinel}{correlation}(
  s_id,
  a_id,
  id,
  group,
  value,
  data,
  type = "correlation",
  n_min,
  output = "overall",
  ...
)
}
\arguments{
\item{s_id}{A character vector of unique sentinel station identifiers. If
this variable is not  specified, the \code{sentinel()} function defaults to
\code{s_id} equaling all unique values in \code{id}.}

\item{a_id}{A character vector of unique auxiliary station identifiers. If
this variable is not specified, the \code{sentinel()} function defaults to
\code{a_id} equaling all unique values in \code{id}.}

\item{id}{A character vector matching the name of the variable in \code{data}
containing station identifiers}

\item{group}{A character vector matching the name of the variable in \code{data}
containing grouping information for which to pair sentinel and auxiliary
stations.}

\item{value}{A character vector matching the name of the variable in \code{data}
containing the response variable for which statistics are computed.}

\item{data}{A data.frame, tibble, sp, or sf object containing the \code{id},
\code{group}, and \code{value} variables.}

\item{type}{The type of summary -- fixed at "correlation"}

\item{n_min}{An integer specifying the minimum number of matching \code{group}
variables in each sentinel-auxiliary pair required to compute relevant
statistics}

\item{output}{A character vector indicating the desired output. \code{output}
can have three values: "overall", "individual", and "dataset". If "overall
is in \code{output}, overall summary statistics are returned for each sentinel
station. If "individual" is in \code{output}, statistics are returned for each
sentinel-auxiliary station pair. If "dataset" is in \code{output}, a data.frame
is returned with all data used to compute relevant statistics. Defaults to
"overall".}

\item{...}{Additional arguments corresponding to arguments from the \code{cor.test}
function in the \code{stats} package (see \code{?stats::cor.test} for
more information).}
}
\value{
A list containing sublists for the corresponding \code{output} specified
}
\description{
\code{sentinel()} computes statistics for an arbitrary number of sentinel
stations relative to an arbitrary number of auxiliary stations.
}
