% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.R
\name{sentinel}
\alias{sentinel}
\title{Compute sentinel station statistics}
\usage{
sentinel(
  s_id,
  a_id,
  id,
  group,
  value,
  data,
  n_min,
  type,
  output = "overall",
  ...
)
}
\arguments{
\item{s_id}{A character vector of unique sentinel station identifiers. If
this variable is not  specified, the \code{sentinel()} function defaults to
\code{s_id} equaling all unique values in \code{id}.}

\item{a_id}{A character vector of unique auxiliary station identifiers. If
this variable is not specified, the \code{sentinel()} function defaults to
\code{a_id} equaling all unique values in \code{id}.}

\item{id}{A character vector matching the name of the variable in \code{data}
containing station identifiers}

\item{group}{A character vector matching the name of the variable in \code{data}
containing grouping information for which to pair sentinel and auxiliary
stations.}

\item{value}{A character vector matching the name of the variable in \code{data}
containing the response variable for which statistics are computed.}

\item{data}{A data.frame, tibble, sp, or sf object containing the \code{id},
\code{group}, and \code{value} variables.}

\item{n_min}{An integer specifying the minimum number of matching \code{group}
variables in each sentinel-auxiliary pair required to compute relevant
statistics}

\item{type}{A character vector indicating the type of statistic desired. The
value of type uses the appropriate generic function:
\itemize{
\item \code{type = "correlation": sentinel.correlation}
} For more information, see \code{?sentinel.type}, where type matches the
character string normally specified by \code{type} (for example, see
\code{?sentinel.correlation} for information about the function when
\code{type = "correlation"})}

\item{output}{A character vector indicating the desired output. \code{output}
can have three values: \code{"overall"}, \code{"individual"}, and \code{"dataset"}. If \code{"overall"}
is in \code{output}, overall summary statistics are returned for each sentinel
station. If \code{"individual"} is in \code{output}, statistics are returned for each
sentinel-auxiliary station pair. If \code{"dataset"} is in \code{output}, a data.frame
is returned with all data used to compute relevant statistics. Defaults to
\code{"overall"}.}

\item{...}{Additional arguments provided to sentinel that correspond to the
value of the \code{type} argument.}
}
\value{
A list containing sublists for the corresponding \code{output} specified
}
\description{
\code{sentinel()} computes statistics for an arbitrary number of sentinel
stations relative to an arbitrary number of auxiliary stations.
}
\examples{
id <-  c(rep("Station1", 10), rep("Station2", 8), rep("Station3", 9))
data <- data.frame(id = c(rep("Station1", 10), rep("Station2", 8), rep("Station3", 9)),
group = c(1:10, 1:8, 1:9), value = rnorm(10 + 8 + 9))
output <- sentinel(id = "id",
group = "group", value = "value", data = data, type = "correlation",
n_min = 0, output = c("overall", "individual", "dataset"), method = "spearman")
}
