% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stationdata}
\alias{stationdata}
\title{Paired station data}
\format{
A data frame with 116 rows and 5 variables:
\describe{
\item{stations}{A character vector of station identifiers}
\item{year}{A numeric vector containing year}
\item{response}{A numeric vector of response values}
\item{xcoord}{A numeric vector of x coordinates}
\item{ycoord}{A numeric vector of y coordinates}
}
}
\usage{
stationdata
}
\description{
A dataset containing stations, auxiliary information, and a response.
}
\keyword{datasets}
